修复cron初始化时pid不存在的错误

+++ core/Group/Cron/Cron.php
@@ -81,7 +81,7 @@ class Cron
         if (file_exists("runtime/pid"))
         $pid = file_get_contents("runtime/pid");
 
-        if ($pid) {
+        if (!empty($pid) && $pid) {
             if (swoole_process::kill($pid, 0)) {
                 swoole_process::kill($pid, SIGUSR1);
                 $filesystem = new \Filesystem();
